Use list comp in response parsing (instead of map)
Users of this library will expect to be able to index into the
returned posts, so we really don't want to return a map object
Tempting though it may be to just fix the line in the test code that
threw the error, this really should be a list.

I elected to use a list comp instead of list(map( because that makes
it fit with the rest of the parsing code (which use list and dict
comprehensions).